refactor(conf.js): centralização de arquivos de apoio em 'utils'
