+++ src/services/auth-service.js
@@ -2,7 +2,7 @@
 
 const jwt = require('jsonwebtoken')
 
-const { conf } = require('../conf')
+const { conf } = require('../utils/conf')
 
 const PRIVATE_KEY = 'f5b99242-6504-4ca3-90f2-05e78e5761ef'
